doom_pipeline/src/requirements.txt
Pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0